requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
channels>=4.0.0
daphne>=4.0.0
psutil>=5.9.0
Pillow>=9.5.0
django-filter>=23.2
django-cors-headers>=4.0.0
python-dotenv>=1.0.0 